6月1日： 1、修改了release的工程报错。 2、修复了当网络不通畅的情况下，进入学习详情造成的无法退出的bug。 3、修复了学习详情中关于评论和赞显示的bug。 4、修复了出现大量评论时，删除评论按钮消失的bug。 5、修复了关于我们和意见反馈web页面显示错位的bug。 6、修复了老师点击学生成绩单造成崩溃的bug。

+++ ChuangYeBa/ArticleInfoDAO.m
@@ -45,7 +45,7 @@ static ArticleInfoDAO *sharedManager = nil;
 
 - (NSString *)applicationDocumentsDirectoryFile:(NSInteger)tag {
     NSString *documentDirectory = [NSSearchPathForDirectoriesInDomains(NSDocumentDirectory, NSUserDomainMask, YES) lastObject];
-    NSString *fileNameWithTag = [FILE_NAME stringByAppendingString:[NSString stringWithFormat:@"%u",tag]];
+    NSString *fileNameWithTag = [FILE_NAME stringByAppendingString:[NSString stringWithFormat:@"%ld",(long)tag]];
     NSString *path = [documentDirectory stringByAppendingPathComponent:fileNameWithTag];
     return path;
 }
